Implement routers for ws and wf
